requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.0.26
python-dotenv==1.0.1